pyproject.toml
[project]
name = "vk-web"
version = "0.1.0"
description = ""
authors = [
    {name = "Artemiy",email = "artemsikor18@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "jinja2 (>=3.1.6,<4.0.0)",
    "fastapi (>=0.119.0,<0.120.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "cachetools (>=6.2.2,<7.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "aiobotocore (>=2.26.0,<3.0.0)",
    "minio (>=7.2.20,<8.0.0)",
    "waitress (>=3.0.2,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.scripts]
app = 'py main.py'
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
